Fix double addition of transaction amount
Line #195 was already adding the transaction to paymentsDone, hence line #200 was again trying to add it and save it to the object.

To prevent any errors, I've used instead Parse built-in increment() function to make sure we just increment the value that the object already has.

TODO: remove the calculation of the 'paymentsDone' but rely on currentTotal property.